Completing the integration tests to verify the default JpaRepository & JpaSpecificationExecutor methods to be complaint with the Multi Tenancy implementation. Custom Repository methods like FindAllByName() are automatically using CriteriaBuilder, so these are also compliant. Adding Jpa Model gen for building typesafe Specification functions
