+++ pom.xml
@@ -24,6 +24,7 @@
         <maven.surefire.plugin.version>3.0.0-M4</maven.surefire.plugin.version>
         <test-container-postgres>1.14.3</test-container-postgres>
         <apache-collections>4.4</apache-collections>
+        <jpa-model-gen>5.4.22.Final</jpa-model-gen>
     </properties>
 
     <dependencies>
@@ -81,6 +82,12 @@
             <artifactId>lombok</artifactId>
             <scope>provided</scope>
             <optional>true</optional>
+        </dependency>
+        <dependency>
+            <groupId>org.hibernate</groupId>
+            <artifactId>hibernate-jpamodelgen</artifactId>
+            <version>${jpa-model-gen}</version>
+            <scope>provided</scope>
         </dependency>
         <dependency>
             <groupId>org.springframework.boot</groupId>
