+++ src/main/java/com/github/mdevloo/multi/tenancy/fwk/multitenancy/UnknownTenantException.java
@@ -1,6 +1,6 @@
 package com.github.mdevloo.multi.tenancy.fwk.multitenancy;
 
-public class UnknownTenantException extends RuntimeException {
+public final class UnknownTenantException extends RuntimeException {
 
   public UnknownTenantException(final String message) {
     super(message);
